requirements.txt
flask~=1.1.2
requests~=2.24.0
xmltodict~=0.12.0
pillow~=8.0.1